#37287 #37353 (2.2.0.89) FEAT_SUBTYPE, FEAT_PRIORITY 型を定義し、型の置換を継続中。 / Define FEAT_SUBTYPE, FEAT_PRIORITY ongoing type replacement.

+++ src/main-win.c
@@ -1691,7 +1691,7 @@ static bool init_graphics(void)
 	/* if (can_use_graphics != arg_graphics) */
 	{
 		char buf[1024];
-		int wid, hgt, twid, thgt, ox, oy;
+		BYTE wid, hgt, twid, thgt, ox, oy;
 		cptr name;
 
 		if (arg_graphics == GRAPHICS_ADAM_BOLT)
